+++ packages/js/src/util/assert.ts
@@ -1,6 +1,6 @@
 import { CLIENT_VERSION } from '../env';
 
-export function assert(value: any, message: string, debug: boolean) {
+export function assert(value: unknown, message: string, debug: boolean) {
     if (debug && !value) {
         console.error(`Flare JavaScript client v${CLIENT_VERSION}: ${message}`);
     }