+++ playground/vue/sections/EnrichmentSection.vue
@@ -1,6 +1,5 @@
 <script lang="ts" setup>
 import { flare } from '../../shared/initFlare';
-
 import Button from '../components/Button.vue';
 import TestSection from '../components/TestSection.vue';
 </script>